Migrate victory-zoom-container to TypeScript
