+++ .changeset/silver-weeks-attend.md
@@ -1,0 +1,5 @@
+---
+"victory-zoom-container": patch
+---
+
+Migrate victory-zoom-container to TypeScript
