--- packages/victory-zoom-container/src/index.js
@@ -1,5 +1,0 @@
-export {
-  zoomContainerMixin,
-  default as VictoryZoomContainer,
-} from "./victory-zoom-container";
-export { default as ZoomHelpers, RawZoomHelpers } from "./zoom-helpers";
