+++ packages/victory-zoom-container/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./victory-zoom-container";
+export * from "./zoom-helpers";
